requirements.txt
# TODO add version number


##################################################
############### Standard Framework ###############
##################################################
# prefer do manually
# cython3 --install-option="--no-cython-compile"
pandas
numpy
sklearn
tqdm
overrides
# image libraries
wand
pillow
opencv-python
# OCR liibs
pytesseract
pyocr


##################################################
################## Code Quality ##################
##################################################
flake8


##################################################
############ Deep Learning framework #############
##################################################
# prefer do manually
# .[tensorflow-cpu] --install-option='-e'



##################################################
###################### NLP #######################
##################################################
# NLP
spacy


##################################################
################### Audio Libs ###################
##################################################
# pyAudioAnalysis
# eyed3
# pydub
# madmom
librosa


##################################################
################## Sphinx Docs ###################
##################################################
# sphinx doc
recommonmark
sphinx_rtd_theme
